+++ scripts/content.js
@@ -1,10 +1,42 @@
-// Configuration
+const DISPLAY_NONE = "display: none";
 
 const PAGE_TYPES = {
   HOME: "home",
   SUBSCRIPTIONS: "subscriptions",
   VIDEO: "video",
+  SEARCH: "search",
+  TRENDING: "trending",
 };
+
+function saveSettings() {
+  const a = document.createElement("a");
+  chrome.storage.local.get(
+    ["tubemod_elements", "tubemod_version"],
+    (result) => {
+      if (chrome.runtime.lastError) {
+        console.error(chrome.runtime.lastError);
+      } else {
+        console.log(result);
+        const file = new Blob([JSON.stringify(result)], {
+          type: "application/json",
+        });
+        a.href = URL.createObjectURL(file);
+        a.download = "tubeModSettings.json";
+        a.click();
+      }
+    }
+  );
+  document.body.removeChild(a);
+}
+
+function importSettings(settings) {
+  chrome.storage.local.set({
+    tubemod_elements: JSON.parse(settings)["tubemod_elements"],
+    tubemod_version: JSON.parse(settings)["tubemod_version"],
+  });
+  location.reload();
+  alert("TubeMod settings uploaded and applied!");
+}
 
 function getCurrentPageType() {
   const url = window.location.href;
@@ -17,18 +49,22 @@ function getCurrentPageType() {
     return PAGE_TYPES.VIDEO;
   } else if (url.includes("/feed/subscriptions")) {
     return PAGE_TYPES.SUBSCRIPTIONS;
+  } else if (url.includes("/results?search_query")) {
+    return PAGE_TYPES.SEARCH;
+  } else if (url.includes("/feed/trending")) {
+    return PAGE_TYPES.TRENDING;
   }
   return null;
 }
 
 const STORAGE = {
-  tubemod_version: "1.6.0",
+  tubemod_version: "1.7.0",
   tubemod_elements: [
     {
       id: "scheduled-videos",
       selector: "//ytd-rich-item-renderer[.//ytd-toggle-button-renderer]",
       checked: false,
-      category: "General",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
@@ -36,49 +72,56 @@ const STORAGE = {
       selector:
         "//ytd-rich-item-renderer[.//div[@id='meta']/ytd-badge-supported-renderer[not(@hidden)]]",
       checked: false,
-      category: "General",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "video-previews",
       selector: "//div[@id='video-preview' or @id='mouseover-overlay']",
       checked: false,
-      category: "General",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "video-thumbnails",
       selector: "//ytd-thumbnail | //ytd-playlist-thumbnail",
       checked: false,
-      category: "General",
+      style: DISPLAY_NONE,
+      pageTypes: [],
+    },
+    {
+      id: "video-thumbnails-blur",
+      selector: "//ytd-thumbnail | //ytd-playlist-thumbnail",
+      checked: false,
+      style: "filter: blur(15px); -webkit-filter: blur(15px);",
       pageTypes: [],
     },
     {
       id: "video-meta-data",
       selector: "//div[@id='metadata-line']",
       checked: false,
-      category: "General",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "logo",
       selector: "//ytd-topbar-logo-renderer",
       checked: false,
-      category: "Header",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "search-bar",
       selector: "//ytd-searchbox[@id='search']",
       checked: false,
-      category: "Header",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "microphone-search",
       selector: "//*[@id='voice-search-button']",
       checked: false,
-      category: "Header",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
@@ -86,14 +129,21 @@ const STORAGE = {
       selector:
         "(//div[@id='buttons' and contains(@class, 'ytd-masthead')]//a/../..)[1]",
       checked: false,
-      category: "Header",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "notifications",
       selector: "//ytd-notification-topbar-button-renderer",
       checked: false,
-      category: "Header",
+      style: DISPLAY_NONE,
+      pageTypes: [],
+    },
+    {
+      id: "profile-sign-out",
+      selector: "//ytd-compact-link-renderer/a[@href='/logout']",
+      checked: false,
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
@@ -101,7 +151,7 @@ const STORAGE = {
       selector:
         "//ytd-guide-entry-renderer[a[@href='/']] | //ytd-mini-guide-entry-renderer[a[@href='/']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
@@ -109,7 +159,7 @@ const STORAGE = {
       selector:
         "//ytd-guide-entry-renderer[a[@title='Shorts']] | //ytd-mini-guide-entry-renderer[a[@title='Shorts']] | //ytd-guide-entry-renderer[a[@title='YouTube Shorts']] | //ytd-mini-guide-entry-renderer[a[@title='YouTube Shorts']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
@@ -117,7 +167,7 @@ const STORAGE = {
       selector:
         "//ytd-guide-entry-renderer[a[@href='/feed/subscriptions']] | //ytd-mini-guide-entry-renderer[a[@href='/feed/subscriptions']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
@@ -125,7 +175,7 @@ const STORAGE = {
       selector:
         "(//ytd-guide-entry-renderer[a[@href='https://music.youtube.com/']])[1] | //ytd-mini-guide-entry-renderer[a[@href='https://music.youtube.com/']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
@@ -133,7 +183,7 @@ const STORAGE = {
       selector:
         "(//div[@id='header']/ytd-guide-entry-renderer)[1] | //ytd-mini-guide-entry-renderer[a[@href='/feed/you']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
@@ -141,21 +191,21 @@ const STORAGE = {
       selector:
         "//div[@id='section-items']/ytd-guide-entry-renderer[a[starts-with(@href, '/@')]]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "history",
       selector: "//ytd-guide-entry-renderer[a[@href='/feed/history']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "playlists",
       selector: "//ytd-guide-entry-renderer[a[@href='/feed/playlists']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
@@ -163,7 +213,7 @@ const STORAGE = {
       selector:
         "//ytd-guide-entry-renderer[a[starts-with(@href, 'https://studio.youtube.com/channel')]]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
@@ -171,35 +221,35 @@ const STORAGE = {
       selector:
         "//ytd-guide-entry-renderer[a[@href='/feed/storefront?bp=ogUCKAQ%3D']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "watch-later",
       selector: "//ytd-guide-entry-renderer[a[@href='/playlist?list=WL']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "courses",
       selector: "//ytd-guide-entry-renderer[a[@href='/feed/courses']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "liked-videos",
       selector: "//ytd-guide-entry-renderer[a[@href='/playlist?list=LL']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "my-clips",
       selector: "//ytd-guide-entry-renderer[a[@href='/feed/clips']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
@@ -207,42 +257,42 @@ const STORAGE = {
       selector:
         "//div[@id='section-items']/ytd-guide-downloads-entry-renderer  | //ytd-mini-guide-entry-renderer[a[@href='/feed/downloads']]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "subscriptions-panel",
       selector: "(//ytd-guide-section-renderer)[2]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "explore-panel",
       selector: "(//ytd-guide-section-renderer)[3]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "youtube-panel",
       selector: "(//ytd-guide-section-renderer)[4]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "youtube-settings",
       selector: "(//ytd-guide-section-renderer)[5]",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
       id: "youtube-footer",
       selector: "//ytd-guide-renderer/div[@id='footer']",
       checked: false,
-      category: "Sidebar",
+      style: DISPLAY_NONE,
       pageTypes: [],
     },
     {
@@ -250,7 +300,7 @@ const STORAGE = {
       selector:
         "//ytd-feed-filter-chip-bar-renderer[@component-style='FEED_FILTER_CHIP_BAR_STYLE_TYPE_DEFAULT']/..",
       checked: false,
-      category: "HomePage",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.HOME],
     },
     {
@@ -258,98 +308,134 @@ const STORAGE = {
       selector:
         "//ytd-ad-slot-renderer/ancestor::ytd-rich-item-renderer | //*[@id='player-ads'] | //ytd-banner-promo-renderer/..",
       checked: false,
-      category: "HomePage",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.HOME],
     },
     {
       id: "posts",
       selector: "//ytd-rich-shelf-renderer/../.. | //ytd-rich-shelf-renderer",
       checked: false,
-      category: "HomePage",
+      style: DISPLAY_NONE,
+      pageTypes: [PAGE_TYPES.HOME],
+    },
+    {
+      id: "playlist-mix",
+      selector:
+        "//ytd-rich-item-renderer[.//ytd-playlist-thumbnail[not(@hidden)]]",
+      checked: false,
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.HOME],
     },
     // {
     //   id: "homepage-video-info",
     //   selector: "//div[@id='metadata-line']",
     //   checked: false,
-    //   category: "HomePage",
+    //   style: DISPLAY_NONE,
     //   pageTypes: [PAGE_TYPES.HOME],
     // },
     // {
     //   id: "homepage-views",
     //   selector: "//div[@id='metadata-line']/span[1]",
     //   checked: false,
-    //   category: "HomePage",
+    //   style: DISPLAY_NONE,
     //   pageTypes: [PAGE_TYPES.HOME],
     // },
     // {
     //   id: "time-posted",
     //   selector: "//div[@id='metadata-line']/span[2]",
     //   checked: false,
-    //   category: "HomePage",
+    //   style: DISPLAY_NONE,
     //   pageTypes: [PAGE_TYPES.HOME],
     // },
     {
       id: "subscriptions-shorts",
       selector: "//ytd-rich-shelf-renderer/../.. | //ytd-rich-shelf-renderer",
       checked: false,
-      category: "Subscriptions",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.SUBSCRIPTIONS],
+    },
+    {
+      id: "search-shorts",
+      selector: "//ytd-reel-shelf-renderer",
+      checked: false,
+      style: DISPLAY_NONE,
+      pageTypes: [PAGE_TYPES.SEARCH],
+    },
+    {
+      id: "search-people-also-searched",
+      selector: "//div[@id='contents']/ytd-horizontal-card-list-renderer",
+      checked: false,
+      style: DISPLAY_NONE,
+      pageTypes: [PAGE_TYPES.SEARCH],
+    },
+    {
+      id: "trending-shorts",
+      selector: "//ytd-video-renderer[.//a[contains(@href, '/shorts')]]",
+      checked: false,
+      style: DISPLAY_NONE,
+      pageTypes: [PAGE_TYPES.TRENDING],
     },
     {
       id: "video-title",
       selector: "//div[@id='above-the-fold']/div[@id='title']",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-subscribers",
       selector: "(//div[@id='upload-info'])[1]/yt-formatted-string",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
+      pageTypes: [PAGE_TYPES.VIDEO],
+    },
+    {
+      id: "video-subscribe-button",
+      selector: "//yt-button-shape[@id='subscribe-button-shape']",
+      checked: false,
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-subscribed-button",
       selector: "//div[@id='notification-preference-button']",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-likes-dislikes",
       selector: "//segmented-like-dislike-button-view-model",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-share",
       selector: "//div[@id='top-level-buttons-computed']/yt-button-view-model",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-download",
       selector: "//ytd-download-button-renderer",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-extra-buttons",
       selector: "//div[@id='flexible-item-buttons']/yt-button-view-model",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-description",
       selector: "//div[@id='description-inner']/parent::div",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
@@ -357,7 +443,7 @@ const STORAGE = {
       selector:
         "//div[@id='description']//div[@id='info-container']//yt-formatted-string[@id='info']//span[position()=1 or position()=2]",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
@@ -365,49 +451,56 @@ const STORAGE = {
       selector:
         "//ytd-horizontal-card-list-renderer[contains (@class, 'ytd-structured-description-content-renderer')]",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-transcription-description",
       selector: "//ytd-video-description-transcript-section-renderer",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-channel-links-description",
       selector: "//ytd-video-description-infocards-section-renderer",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
+      pageTypes: [PAGE_TYPES.VIDEO],
+    },
+    {
+      id: "video-live-chat-replay",
+      selector: "//div[@id='teaser-carousel']",
+      checked: false,
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-comments",
       selector: "//ytd-comments[@id='comments']",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-categories-games",
       selector: "//ytd-rich-metadata-row-renderer/../..",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-ads",
       selector: "//div[@id='player-ads']",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-tabs",
       selector: "//yt-related-chip-cloud-renderer",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
@@ -415,21 +508,21 @@ const STORAGE = {
       selector:
         "//div[@id='contents']/parent::ytd-item-section-renderer[contains(@class, 'watch-next')]",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-suggested-shorts",
       selector: "//ytd-reel-shelf-renderer",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
     {
       id: "video-suggestion-wall",
       selector: "//div[@class='ytp-endscreen-content']",
       checked: false,
-      category: "Video",
+      style: DISPLAY_NONE,
       pageTypes: [PAGE_TYPES.VIDEO],
     },
   ],
@@ -513,7 +606,7 @@ class YouTubeElement {
   }
 
   applyVisibility(hide) {
-    const displayValue = hide ? "none" : "";
+    const displayValue = hide ? this.style : "";
     const elements = document.evaluate(
       this.selector,
       document,
@@ -523,7 +616,29 @@ class YouTubeElement {
     );
 
     for (let i = 0; i < elements.snapshotLength; i++) {
-      elements.snapshotItem(i).style.display = displayValue;
+      elements.snapshotItem(i).style = `${displayValue}`;
+    }
+
+    if (this.id === "you") {
+      const elementWithTopBorder = document.querySelector(
+        "ytd-guide-collapsible-section-entry-renderer"
+      );
+      if (elementWithTopBorder) {
+        elementWithTopBorder.style.borderTop = hide
+          ? "none"
+          : "1px solid var(--yt-spec-10-percent-layer)";
+      }
+    }
+
+    if (this.id === "my-clips") {
+      const elementWithBottomBorder = document.querySelector(
+        "ytd-guide-section-renderer"
+      );
+      if (elementWithBottomBorder) {
+        elementWithBottomBorder.style.borderBottom = hide
+          ? "none"
+          : "1px solid var(--yt-spec-10-percent-layer)";
+      }
     }
   }
 }
@@ -643,6 +758,10 @@ class TubeMod {
   handleMessage(request) {
     if (request.action === "clearLocalStorage") {
       this.clearLocalStorage();
+    } else if (request.action === "saveSettings") {
+      saveSettings();
+    } else if (request.action === "importSettings") {
+      importSettings(request.content);
     } else {
       this.elementManager.handleAction(request.action);
     }