+++ documentation/ReleaseNotes.md
@@ -1,6 +1,37 @@
 # TubeMod - Release Notes
 
-## 1.6.0 (Released)
+## 1.7.0
+
+Add:
+
+- General:
+  - Blur Video Thumbnails
+
+- Profile Settings:
+  - Hide 'Sign Out' Button
+
+- Home:
+  - Hide Mix Playlists
+
+- Search:
+  - Hide Shorts
+  - Hide 'People Also Search For'
+
+- Trending:
+  - Hide Shorts
+
+- Video:
+  - Hide Subscribe Button
+  - Hide Video Live Chat Replay
+
+- Import / Export TubeMod Settings
+
+Fix: 
+
+- Toggles now maintain their width regardless of label length 
+- When removing everything from the sidebar, no more horizontal bars are shown
+
+## 1.6.0 
 
 Add:
 
@@ -28,7 +59,7 @@ Fix:
 - Sidebar
   - Fixed an issue with the Sidebar Footer, as it was messing up with posts creation. [View the issue On GitHub](https://github.com/Pedro-Gregorio/TubeMod/issues/60).
 
-## 1.5.0 (Released)
+## 1.5.0 
 
 ### **Note from Pedro**:
 
@@ -74,7 +105,7 @@ Disabled:
 
 Note: Both these features are related to taking actions, and that will be a later focus of the extension. For now, the focus will be around hiding content, as gracefully as possible, without impacting on performance.
 
-## 1.4.0 (Released)
+## 1.4.0 
 
 Add:
 
@@ -82,7 +113,7 @@ Add:
 - Hide Transfers Sidebar Option, for Premium users
 - Hide Video Suggested Shorts Option
 
-## 1.3.0 (Released)
+## 1.3.0 
 
 Add:
 
@@ -95,7 +126,7 @@ Fix:
 
 - Hide Home Page Tabs was affecting Channel Video Section Tabs
 
-## 1.2.0 (Released)
+## 1.2.0 
 
 Add:
 
@@ -105,7 +136,7 @@ Fix:
 
 - "Create Button" Locator for Opera
 
-## 1.1.0 (Released)
+## 1.1.0 
 
 Add:
 
@@ -115,7 +146,7 @@ Change:
 
 - Subscription Page Shorts Locator
 
-## 1.0.0 (Released)
+## 1.0.0 
 
 Add: